requirements.txt
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
flake8==6.0.0
Flask==2.2.3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
outcome==1.2.0
packaging==23.0
pycodestyle==2.10.0
pyflakes==3.0.1
PySocks==1.7.1
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
spotipy==2.22.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
Werkzeug==2.2.3
wsproto==1.2.0
